ViewModel changed to FragmentHome from MainActivity
